backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
razorpay==1.4.1
supabase==2.0.2
python-dotenv==1.0.0
bcrypt==4.2.0
PyJWT==2.9.0
requests==2.32.3
twilio==9.0.0